Move getActiveSnapshotNodes to selectors/index.js
