+++ src/selectors/index.test.js
@@ -3,7 +3,8 @@ import {
   getSnapshotHistory,
   getActiveSnapshotMessage,
   getActiveSnapshotTimestamp,
-  getActiveSchema
+  getActiveSchema,
+  getActiveSnapshotNodes
 } from './index';
 
 describe('Selectors', () => {
@@ -38,6 +39,26 @@ describe('Selectors', () => {
       const schema = getActiveSchema(mockState);
       expect(schema).toEqual(expect.any(String));
       expect(JSON.parse(schema)).toEqual(expect.any(Object));
+    });
+  });
+
+  describe('getActiveSnapshotNodes', () => {
+    it('retrieves a list of nodes for the active snapshot', () => {
+      expect(getActiveSnapshotNodes(mockState)).toEqual(
+        expect.arrayContaining([expect.any(String)])
+      );
+    });
+
+    it('returns an empty array if snapshotNodes is empty', () => {
+      const newMockState = Object.assign({}, mockState, { snapshotNodes: {} });
+      expect(getActiveSnapshotNodes(newMockState)).toEqual([]);
+    });
+
+    it('returns an empty array if activeSnapshot is undefined', () => {
+      const newMockState = Object.assign({}, mockState, {
+        activeSnapshot: undefined
+      });
+      expect(getActiveSnapshotNodes(newMockState)).toEqual([]);
     });
   });
 });
